Progress on block persistence
This rehashes the user password on a successful login, however, it doesn't purge the active sessions since it only stores the user id and accepts it blindly
