+++ components/Authentication.php
@@ -1,9 +1,12 @@
 <?php namespace RainLab\User\Components;
 
 use Cms;
+use Auth;
 use Flash;
 use Config;
+use Request;
 use Cms\Classes\ComponentBase;
+use RainLab\User\Models\Setting as UserSetting;
 use RainLab\User\Helpers\User as UserHelper;
 use NotFoundException;
 
@@ -173,6 +176,10 @@ class Authentication extends ComponentBase
      */
     public function useRememberMe(): bool
     {
+        if ($this->showRememberMe()) {
+            return (bool) input('remember');
+        }
+
         return $this->property('rememberMe') !== self::REMEMBER_NEVER;
     }
 
@@ -190,5 +197,29 @@ class Authentication extends ComponentBase
     public function showUsernameField()
     {
         return UserHelper::showUsername();
+    }
+
+    /**
+     * prepareOtherUserSessions will log out other sessions, if set by configuration
+     */
+    protected function prepareOtherUserSessions()
+    {
+        if (!UserSetting::get('block_persistence', false)) {
+            return;
+        }
+
+        if ($password = post('password')) {
+            Auth::logoutOtherDevices($password);
+        }
+    }
+
+    /**
+     * prepareAuthenticatedSession
+     */
+    protected function prepareAuthenticatedSession()
+    {
+        if (Request::hasSession()) {
+            Request::session()->regenerate();
+        }
     }
 }
