+++ components/authentication/ActionTwoFactorLogin.php
@@ -45,6 +45,10 @@ trait ActionTwoFactorLogin
             return $this->actionLogin();
         }
 
+        // Needed for Auth::logoutOtherDevices
+        Auth::setUser($user);
+        $this->prepareOtherUserSessions();
+
         Session::put('login.id', $user->getKey());
         Session::put('login.remember', $this->useRememberMe());
 
@@ -69,7 +73,7 @@ trait ActionTwoFactorLogin
 
         Auth::login($user, $this->useRememberMe());
 
-        Request::session()->regenerate();
+        $this->prepareAuthenticatedSession();
     }
 
     /**
